make a particle framework
This is taking longer than expected to make a particle demo that won't
fall apart as soon as I try to give it some form of useful structure.
If I were making a simple particle emitter that had its own "ammo box",
so to speak, and it knew where its boundaries were and could update the
particles on its own, then it would be simple.  But I am making it
possible to use more than one particle emitter in a single region, and
that region may be a sphere or polygon.  These seemingly simple
requirements made the design much more complex.