particles regions and emitters now work on interfaces; finished particle updater
I decided to give the particle emitter point and bar and particle
regions for circle and polygon their own files.  They now inherit from
indices, which made the Particle emitter's job easier.  Also added a
polygon to the PrimitiveGeneration.  This will allow me to create a
renderable shape that matches that of a particle polygon region, giving
the illusion that the particles are confined to the walls.

I intend to draw particles and geometry at the same time, so I will need
some sort of shader "manager", though I'd rather find a more OO approach
than simply a "manager".