Renaming project
I realized as I was working on this project that it incorporates ideas
that are not "barebones" but that are necessary to figure out how
particle regions and emitters work and can be organized, so I copied
what I had into a new repo, render_particles_2D_basic.  I will make a
much more simple particle system there.