FreeType stuff; particle regions and emitters can now set a transform
FreeType stuff now correctly uses texture units and texture samplers
instead of relying on the default 0 unit and 0 sampler.  Updated the
particle regions and particle emitters to take a transform and modify
their position and direction info accordingly.  This eliminates the need
to calculate the info at program start and then be stuck with it.